Scheme 48 0.46 from tarball.

+++ c/unix/io.c
@@ -71,7 +71,7 @@ ps_read_char(FILE *port, bool *eofp, int *status, bool peekp)
 }
 
 long
-ps_read_integer(FILE *port, bool *eofp, int *status)
+ps_read_integer(FILE *port, bool *eofp, long *status)
 {
   long result;
   int ch;
